Extract session and SSE handling
